[lib] Use uint8_t instead of size_t in memmove

+++ lib/stdlib/string.c
@@ -1,4 +1,5 @@
 #include <stddef.h>
+#include <stdint.h>
 
 #include "string.h"
 
@@ -16,13 +17,13 @@ size_t strlen(const char *string)
 void *memmove(void *dest, const void *src, size_t n)
 {
     // FIXME: could these casts possibly cause issues in the future?
-    size_t *tmp_dest = (size_t *)dest;
-    size_t *tmp_src = (size_t *)src;
+    uint8_t *tmp_dest = (uint8_t *)dest;
+    uint8_t *tmp_src = (uint8_t *)src;
 
     // If dest and src are overlapping, directly overwriting dest could
     // make src possibly lose important data that it shouldn't lose yet
     // hence why use the tmp buffer.
-    size_t tmp[n];
+    uint8_t tmp[n];
 
     for (size_t i = 0; i < n; ++i)
     {
